[32기 임종혁]Feed기능  component화 완료
